+++ src/OpenRiaServices.DomainServices.Client.Web/Framework/TypeForwarding.cs
@@ -1,7 +1,0 @@
-﻿using System.Runtime.CompilerServices;
-
-[assembly: TypeForwardedTo(typeof(OpenRiaServices.DomainServices.Client.ApplicationServices.AuthenticationDomainContextBase))]
-[assembly: TypeForwardedTo(typeof(OpenRiaServices.DomainServices.Client.ApplicationServices.WebContextBase))]
-[assembly: TypeForwardedTo(typeof(OpenRiaServices.DomainServices.Client.ApplicationServices.WebAuthenticationService))]
-[assembly: TypeForwardedTo(typeof(OpenRiaServices.DomainServices.Client.ApplicationServices.WindowsAuthentication))]
-[assembly: TypeForwardedTo(typeof(OpenRiaServices.DomainServices.Client.ApplicationServices.FormsAuthentication))]